ga_ua_extract_data/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   flask
    #   functions-framework
cloudevents==1.9.0
    # via functions-framework
deprecation==2.1.0
    # via cloudevents
flask==2.2.3
    # via functions-framework
functions-framework==3.3.0
    # via -r requirements.in
ga-ua-api @ git+https://github.com/jallen13/ga-ua-api.git
    # via -r requirements.in
gaapi4py @ git+https://github.com/jallen13/gaapi4py.git
    # via ga-ua-api
google-api-core==2.11.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.86.0
    # via gaapi4py
google-auth==2.17.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-storage==2.8.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-storage
googleapis-common-protos==1.59.0
    # via google-api-core
gunicorn==20.1.0
    # via functions-framework
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
numpy==1.24.3
    # via pandas
oauth2client==4.1.3
    # via gaapi4py
packaging==23.1
    # via deprecation
pandas==2.0.1
    # via
    #   -r requirements.in
    #   ga-ua-api
    #   gaapi4py
protobuf==4.22.3
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   google-auth
    #   oauth2client
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
requests==2.28.2
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   python-dateutil
tzdata==2023.3
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.15
    # via requests
watchdog==3.0.0
    # via functions-framework
werkzeug==2.2.3
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
